feat: Implement BotRunnerMaestro for bot execution with logging and SharePoint integration
- Added BotRunnerMaestro class to handle bot execution, logging, and resource monitoring.
- Integrated SharePoint API for uploading logs and managing folders.
- Configured SQL database logging for execution details.
- Created centralized configuration settings using Pydantic.
- Developed LoggerConfig for structured logging with rotation.
- Introduced enumerations for department names and recurrence types.
- Added SQL query for logging execution details.
- Implemented SQLDatabaseConnectorDict for database interactions.
