+++ bot.py
@@ -1,9 +1,9 @@
 import argparse
 import os
 
-from botcity.runner.botcity_local import BotRunnerLocal
-from botcity.runner.botcity_maestro import BotRunnerMaestro
-from botcity.settings.base_settings import settings
+from botcity.botcity_local import BotRunnerLocal
+from botcity.botcity_maestro import BotRunnerMaestro
+from botcity.core.config import settings
 
 
 def parse_args() -> argparse.Namespace:
