+++ botcity/enums/department.py
