+++ botcity/query/insert_log.sql
